iSpeak2/requirements.txt
Flask
torch
torchaudio
git+https://github.com/speechbrain/speechbrain.git@develop
gTTS
transformers
scipy
pandas
nltk
tacotron2
python-docx
speechbrain==0.5.16